+++ tests/tinytest/test_across_lazy.R
@@ -260,7 +260,7 @@ expect_equal_lazy(
     )
   ),
   summarize(test_grp, mpg_my_mean = mean(mpg), mpg_my_median = median(mpg)) |>
-    to_r()
+    as.data.frame()
 )