+++ tests/tinytest/test_arrange.R
@@ -34,7 +34,7 @@ expect_equal(
 expect_equal(
   arrange(test, x1, -x2) |>
     select(starts_with("x")) |>
-    to_r(),
+    as.data.frame(),
   data.frame(
     x1 = c("a", "a", "a", "b", "c"),
     x2 = c(3, 2, 1, 5, 1)