+++ tests/tinytest/test_pull.R
@@ -1,0 +1,19 @@
+source("helpers.R")
+using("tidypolars")
+
+test <- polars::pl$DataFrame(mtcars)
+
+expect_equal(
+  pull(test, mpg),
+  mtcars$mpg
+)
+
+expect_equal(
+  pull(test, "mpg"),
+  mtcars$mpg
+)
+
+expect_equal(
+  pull(test, 1),
+  mtcars$mpg
+)