+++ NEWS.md
@@ -13,6 +13,8 @@
 
 * Add support for argument `.keep` in `mutate()` (#80).
 
+* *Experimental* support of `rowwise()` (#40).
+
 
 # tidypolars 0.5.0