+++ NEWS.md
@@ -17,7 +17,9 @@
 
 * Add support for `group_vars()` and `group_keys()` (#81).
 
-* *Experimental* support of `rowwise()` (#40).
+* **Experimental** support of `rowwise()`. For now, this is limited to a few 
+  functions: `mean()`, `min()`, `max()`, `sum()`, `all()`, `any()`. `rowwise()`
+  and `group_by()` cannot be used at the same time (#40).
 
 # tidypolars 0.5.0