+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidypolars (development version)
+
 # tidypolars 0.6.0
 
 `tidypolars` requires `polars` >= 0.15.0.